Configurey changes for MinGW.
Updated to use armeabi-v7a, android-10, set to CONFIGURE_QT=1
Fix for getQMakeConf when it returns unexpanded QMake strings (e.g. $$ANDROID_NDK_TOOLCHAIN/bin). configure replaces $$ with $ and replaces the var with the result of running eval on this.
Use Makefile.win32-g++-sh to build QMake on MSYS.