Qt Android BroadcastReceiver wapper.
Android BroadcastReceiver Java class into main Jar file to be loaded by Ministro.

Although it is very generic, C++ implementation is currently in Connectivity
module until a good way of making it public in main Qt repository is found.

I moved this file to mobility package, because I'm planning to split the
mobility and main packages into separate .jar files, and because this is used
only by mobility.

Many thanks to Lauri Laanmets <lauri.laanmets@proekspert.ee> for this work